+++ R/methods-fragments.R
@@ -3,7 +3,7 @@ setMethod("calculateFragments", c("character", "missing"),
                    modifications = c(C = 57.02146),
                    neutralLoss = defaultNeutralLoss(),
                    verbose = isMSnbaseVerbose()) {
-            l <- lapply(sequence, .calculateFragments,
+            l <- lapply(sequence, PSMatch:::.calculateFragments,
                         type = type, z = z, modifications = modifications,
                         neutralLoss = neutralLoss, verbose = verbose)
             return(do.call(rbind, l))